requirements.txt
astroid==2.2.5
Django==2.2.10
gunicorn==19.9.0
requests
django-oauth-toolkit==1.2.0
django-rest-framework==0.1.0
Pillow==7.0.0